Python/frameworks/django/pyproject.toml
[tool.poetry]
name = "eskd_base"
version = "0.1.0"
description = ""
authors = ["Daniil Zaytsev <zdanil34@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.4"
django-debug-toolbar = "^3.7.0"
pydantic = "^1.10.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.group.dev.dependencies]
mypy = "^0.971"
requests = "^2.28.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
